Monte Carlo with Geometric Brownian Motion
An alternative version to perform Monte Carlo Simulations. Geometric Brownian Motion (GBM) is the differential equation assumed in the Black-Scholes options pricing model. Using the Geometric Brownian Motion, we are able to fit a continuous function for price, and therefore, simulate price ahead in time. The assumptions of GBM, in my opinion, are more robust than that of Monte Carlo Simulations using the sole assumption that returns are normally distributed.